requirements.txt
pysnc
discord.py
aiohttp
asyncio
discord-py-slash-command
